Protobuf Replacement (#54)
* Replace logcache protobufs with go-log-cache

Pulls in compiled protobufs from go-log-cache instead of the
deprecated log-cache. Still some references to log-cache in the code,
the main focus was on the protobuf files. Deletes compiled protobufs in
here as they were unused.

Also replaces github/protobuf (deprecated) with golang/protobuf,
loggregator with loggreator v8, grpc-gateway v1 with grpc-gateway v2

* vendor files

* Delete copy of go-log-cache that's pasted here

Delete the copy and pasted code and update the references to point to
go-log-cache directly instead of the copy of the code

* Update go-log-cache reference to master

Previously, the go-log-cache dependency in go.mod was referencing to a feature branch.
Since the feature branch has been merged, the code now points to the
master branch.

code.cloudfoundry.org/go-log-cache/rpc/logcache_v1

Co-authored-by: Matthew Kocher <mkocher@vmware.com>
Co-authored-by: Carson Long <lcarson@vmware.com>