Fixes concurrency bug in store test suite
This commit changes the Store creation logic in the tests so that the
tests will never have multiple Stores accessing the same Metrics
Registry at the same time.

Prior to this commit, some tests (but most notably the "truncates older
envelopes when max size is reached" test) will -when waiting to data
values out of the Metrics Registry- will sometimes get values for the
wrong Registry.

Signed-off-by: Matthew Kocher <mkocher@vmware.com>