Add optional mTLS for the log-cache-syslog-server
Activates mutual TLS in the log-cache-syslog-server. All of the incoming
syslog connections will be verified, if a syslog client CA certificate
is configured. The goal is to harden the communication between the
Syslog Agents running on all BOSH VMs and the Log-Cache Syslog Server
and secure it with mutual TLS.

Changes:
- adjust BOSH job templates and add a field for syslog client CA
- configure the log-cache-syslog-server to use mTLS in case a the syslog
client CA is configured
- add test for log-cache-syslog-server mTLS configuration
- update README.md add paragraph about TLS and mTLS configuration for
the log-cache-syslog-server