Make memory_limit_bytes available as optional bosh property
The code allows both a configurable % and byte limit for memory. Specifically the byte limit makes sense when you want
to run log-cache on a really large VM, colocated with other things, but you don't want it to take up too much
of the VM's memory.

This optional property will do nothing until set, at which point it will override the memory % limit, as documented.